+++ subsys/net/ip/CMakeLists.txt
@@ -59,14 +59,10 @@ if(CONFIG_NET_TCP_ISN_RFC6528)
 endif()
 endif()
 
+# To get private includes like net_shell.h
+zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/lib)
+
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
   ${ARCH_DIR}/${ARCH}/include
   )
-
-if(CONFIG_NET_SHELL)
-zephyr_library_sources(net_shell.c)
-zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/l2)
-zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/lib)
-zephyr_library_link_libraries_ifdef(CONFIG_MBEDTLS mbedTLS)
-endif()
