+++ subsys/net/lib/CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory_ifdef(CONFIG_NET_SOCKETS            sockets)
 add_subdirectory_ifdef(CONFIG_TLS_CREDENTIALS        tls_credentials)
 add_subdirectory_ifdef(CONFIG_NET_CAPTURE            capture)
 add_subdirectory_ifdef(CONFIG_NET_ZPERF              zperf)
+add_subdirectory_ifdef(CONFIG_NET_SHELL              shell)
 
 if (CONFIG_DNS_RESOLVER
     OR CONFIG_MDNS_RESPONDER
